==> Analysis/Annotation/2.annotation.Stats.V2.ipynb <==
# QC & exploration of final curated Erdman genome

from Bio import SeqIO
# from io import StringIO
# from Bio.SeqFeature import SeqFeature, FeatureLocation
import pandas as pd
import bioframe as bf
from Bio.Seq import Seq
from Bio.SeqUtils import GC

df_annotations = pd.read_csv('./CuratedAnnotationV2/Erdman.curated.V2.tsv', sep='\t')
df_annotations.shape

df_annotations.head()

df_annotations["source"].value_counts()

df_annotations.query("type == 'CDS'")["source"].value_counts()

df_annotations["type"].value_counts()

df_annotations["type"].value_counts()

!ls -1 ../../References/H37Rv_Anno_EMBL

records = [record for record in SeqIO.parse('../../References/H37Rv_Anno_EMBL/GCF_000195955.2_ASM19595v2_genomic.embl', 'embl')]    

Rv_Seq = records[0].seq

records = [record for record in SeqIO.parse('./CuratedAnnotationV2/Erdman.curated.V2.gb', 'genbank')]

Erdman_Seq = records[0].seq

len(Rv_Seq)

len(Erdman_Seq)

# Calculate GC content
gc_content = GC(Erdman_Seq)
print(f"Erdman V2 - GC content: {gc_content:.2f}%")

# Calculate GC content
gc_content = GC(Rv_Seq)
print(f"H37Rv - GC content: {gc_content:.2f}%")

# Calculate GC content
gc_content = GC(Erdman_Seq)
print(f"Erdman V2 - GC content: {gc_content:.5f}%")

# Calculate GC content
gc_content = GC(Rv_Seq)
print(f"H37Rv - GC content: {gc_content:.5f}%")
